+++ qiskit/backends/ibmq/ibmqbackend.py
@@ -26,15 +26,17 @@ class IBMQBackend(BaseBackend):
     """Backend class interfacing with the Quantum Experience remotely.
     """
 
-    def __init__(self, configuration, api=None):
+    def __init__(self, configuration, provider, credentials, api):
         """Initialize remote backend for IBM Quantum Experience.
 
         Args:
             configuration (dict): configuration of backend.
+            provider (IBMQProvider): provider.
+            credentials (Credentials): credentials.
             api (IBMQuantumExperience.IBMQuantumExperience.IBMQuantumExperience):
                 api for communicating with the Quantum Experience.
         """
-        super().__init__(configuration=configuration)
+        super().__init__(provider=provider, configuration=configuration)
         self._api = api
         if self._configuration:
             configuration_edit = {}
@@ -46,6 +48,11 @@ class IBMQBackend(BaseBackend):
             # local : False is added to the online device
             self._configuration['local'] = False
 
+        self._credentials = credentials
+        self.hub = credentials.hub
+        self.group = credentials.group
+        self.project = credentials.project
+
     def run(self, qobj):
         """Run qobj asynchronously.
 
@@ -56,7 +63,7 @@ class IBMQBackend(BaseBackend):
             IBMQJob: an instance derived from BaseJob
         """
         job_class = _job_class_from_backend_support(self)
-        job = job_class(self._api, not self.configuration()['simulator'], qobj=qobj)
+        job = job_class(self, None, self._api, not self.configuration()['simulator'], qobj=qobj)
         job.submit()
         return job
 
@@ -247,9 +254,7 @@ class IBMQBackend(BaseBackend):
         for job_info in job_info_list:
             job_class = _job_class_from_job_response(job_info)
             is_device = not bool(self._configuration.get('simulator'))
-            job = job_class(self._api, is_device,
-                            job_id=job_info.get('id'),
-                            backend_name=backend_name,
+            job = job_class(self, job_info.get('id'), self._api, is_device,
                             creation_date=job_info.get('creationDate'))
             job_list.append(job)
         return job_list
@@ -276,11 +281,17 @@ class IBMQBackend(BaseBackend):
                                    .format(job_id, str(ex)))
         job_class = _job_class_from_job_response(job_info)
         is_device = not bool(self._configuration.get('simulator'))
-        job = job_class(self._api, is_device,
-                        job_id=job_info.get('id'),
-                        backend_name=job_info.get('backend').get('name'),
+        job = job_class(self, job_info.get('id'), self._api, is_device,
                         creation_date=job_info.get('creationDate'))
         return job
+
+    def __repr__(self):
+        credentials_info = ''
+        if self.hub:
+            credentials_info = '{}, {}, {}'.format(self.hub, self.group,
+                                                   self.project)
+        return "<{}('{}') from IBMQ({})>".format(
+            self.__class__.__name__, self.name(), credentials_info)
 
 
 class IBMQBackendError(QISKitError):