+++ qiskit/backends/ibmq/ibmqprovider.py
@@ -5,59 +5,73 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""Provider for remote IbmQ backends."""
-from IBMQuantumExperience import IBMQuantumExperience
+"""Provider for remote IBMQ backends with admin features."""
 
-from qiskit._util import _camel_case_to_snake_case
-from qiskit.backends.baseprovider import BaseProvider
-from qiskit.backends.ibmq.ibmqbackend import IBMQBackend
-from qiskit._util import _parse_ibmq_credentials
+from collections import OrderedDict
+
+from qiskit.backends import BaseProvider
+
+from .credentials._configrc import remove_credentials
+from .credentials import (Credentials,
+                          read_credentials_from_qiskitrc, store_credentials, discover_credentials)
+from .ibmqaccounterror import IBMQAccountError
+from .ibmqsingleprovider import IBMQSingleProvider
+
+QE_URL = 'https://quantumexperience.ng.bluemix.net/api'
 
 
 class IBMQProvider(BaseProvider):
-    """Provider for remote IbmQ backends."""
-    def __init__(self, token, url='https://quantumexperience.ng.bluemix.net/api',
-                 hub=None, group=None, project=None, proxies=None, verify=True):
+    """Provider for remote IBMQ backends with admin features.
+
+    This class is the entry point for handling backends from IBMQ, allowing
+    using different accounts.
+    """
+    def __init__(self):
         super().__init__()
 
-        url = _parse_ibmq_credentials(url, hub, group, project)
+        # dict[credentials_unique_id: IBMQSingleProvider]
+        # This attribute stores a reference to the different accounts. The
+        # keys are tuples (hub, group, project), as the convention is that
+        # that tuple uniquely identifies a set of credentials.
+        self._accounts = OrderedDict()
 
-        # Get a connection to IBMQuantumExperience.
-        self._api = self._authenticate(token, url, proxies=proxies, verify=verify)
-
-        # Populate the list of remote backends.
-        self.backends = self._discover_remote_backends()
-
-        # authentication attributes, which uniquely identify the provider instance
-        self._token = token
-        self._url = url
-        self._proxies = proxies
-        self._verify = verify
-
-    def get_backend(self, name):
-        return self.backends[name]
-
-    def available_backends(self):
-        """Get a list of available backends from the IBMQ provider.
-
-        Returns:
-            list[IBMQBackend]: a list of backend instances available
-            from the IBMQ provider.
-        """
+    def backends(self, name=None, filters=None, **kwargs):
         # pylint: disable=arguments-differ
-        return list(self.backends.values())
 
-    def grouped_backend_names(self):
-        return {}
+        # Special handling of the credentials filters.
+        credentials_filter = {}
+        for key in ['token', 'url', 'hub', 'group', 'project']:
+            if key in kwargs:
+                credentials_filter[key] = kwargs.pop(key)
+        providers = [provider for provider in self._accounts.values() if
+                     self._match_all(provider.credentials, credentials_filter)]
 
-    def deprecated_backend_names(self):
+        # Special handling of the `name` parameter, to support alias resolution.
+        if name:
+            aliases = self.aliased_backend_names()
+            aliases.update(self.deprecated_backend_names())
+            name = aliases.get(name, name)
+
+        # Aggregate the list of filtered backends.
+        backends = []
+        for provider in providers:
+            backends = backends + provider.backends(
+                name=name, filters=filters, **kwargs)
+
+        return backends
+
+    @staticmethod
+    def deprecated_backend_names():
+        """Returns deprecated backend names."""
         return {
             'ibmqx_qasm_simulator': 'ibmq_qasm_simulator',
             'ibmqx_hpc_qasm_simulator': 'ibmq_qasm_simulator',
             'real': 'ibmqx1'
             }
 
-    def aliased_backend_names(self):
+    @staticmethod
+    def aliased_backend_names():
+        """Returns aliased backend names."""
         return {
             'ibmq_5_yorktown': 'ibmqx2',
             'ibmq_5_tenerife': 'ibmqx4',
@@ -65,76 +79,165 @@ class IBMQProvider(BaseProvider):
             'ibmq_20_austin': 'QS1_1'
             }
 
-    @classmethod
-    def _authenticate(cls, token, url, proxies=None, verify=True):
-        """
-        Authenticate against the IBMQuantumExperience API.
+    def add_account(self, token, url=QE_URL, **kwargs):
+        """Authenticate against IBMQ and store the account for future use.
 
-        Returns:
-            IBMQuantumExperience.IBMQuantumExperience.IBMQuantumExperience:
-                instance of the IBMQuantumExperience API.
-        Raises:
-            ConnectionError: if the authentication resulted in error.
-        """
-        try:
-            config_dict = {
-                'url': url,
-            }
-            if proxies:
-                config_dict['proxies'] = proxies
-            return IBMQuantumExperience(token, config_dict, verify)
-        except Exception as ex:
-            root_exception = ex
-            if 'License required' in str(ex):
-                # For the 401 License required exception from the API, be
-                # less verbose with the exceptions.
-                root_exception = None
-            raise ConnectionError("Couldn't connect to IBMQuantumExperience server: {0}"
-                                  .format(ex)) from root_exception
-
-    @classmethod
-    def _parse_backend_configuration(cls, config):
-        """
-        Parse a backend configuration returned by IBMQuantumConfiguration.
+        Login into Quantum Experience or IBMQ using the provided credentials,
+        adding the account to the current session. The account is stored in
+        disk for future use.
 
         Args:
-            config (dict): raw configuration as returned by
-                IBMQuantumExperience.
+            token (str): Quantum Experience or IBM Q API token.
+            url (str): URL for Quantum Experience or IBM Q (for IBM Q,
+                including the hub, group and project in the URL).
+            **kwargs (dict):
+                * proxies (dict): Proxy configuration for the API.
+                * verify (bool): If False, ignores SSL certificates errors
+
+        Raises:
+            IBMQAccountError: if the credentials are already in use.
+        """
+        credentials = Credentials(token, url, **kwargs)
+
+        # Check if duplicated credentials are already stored. By convention,
+        # we assume (hub, group, project) is always unique.
+        stored_credentials = read_credentials_from_qiskitrc()
+
+        if credentials.unique_id() in stored_credentials.keys():
+            raise IBMQAccountError('Credentials are already stored')
+
+        self._append_account(credentials)
+
+        # Store the credentials back to disk.
+        store_credentials(credentials)
+
+    def remove_account(self, token, url=QE_URL, **kwargs):
+        """Remove an account from the session and from disk.
+
+        Args:
+            token (str): Quantum Experience or IBM Q API token.
+            url (str): URL for Quantum Experience or IBM Q (for IBM Q,
+                including the hub, group and project in the URL).
+            **kwargs (dict):
+                * proxies (dict): Proxy configuration for the API.
+                * verify (bool): If False, ignores SSL certificates errors
+
+        Raises:
+            IBMQAccountError: if the credentials could not be removed.
+        """
+        removed = False
+        credentials = Credentials(token, url, **kwargs)
+
+        # Check if the credentials are already stored in session or disk. By
+        # convention, we assume (hub, group, project) is always unique.
+        stored_credentials = read_credentials_from_qiskitrc()
+
+        # Try to remove from session.
+        if credentials.unique_id() in self._accounts.keys():
+            del self._accounts[credentials.unique_id()]
+            removed = True
+
+        # Try to remove from disk.
+        if credentials.unique_id() in stored_credentials.keys():
+            remove_credentials(credentials)
+            removed = True
+
+        if not removed:
+            raise IBMQAccountError('Unable to find credentials')
+
+    def remove_accounts(self):
+        """Remove all accounts from this session and optionally from disk."""
+        current_creds = self._accounts.copy()
+        for creds in current_creds:
+            self.remove_account(current_creds[creds].credentials.token,
+                                current_creds[creds].credentials.url)
+
+    def use_account(self, token, url=QE_URL, **kwargs):
+        """Authenticate against IBMQ during this session.
+
+        Login into Quantum Experience or IBMQ using the provided credentials,
+        adding the account to the current session. The account is not stored
+        in disk.
+
+        Args:
+            token (str): Quantum Experience or IBM Q API token.
+            url (str): URL for Quantum Experience or IBM Q (for IBM Q,
+                including the hub, group and project in the URL).
+            **kwargs (dict):
+                * proxies (dict): Proxy configuration for the API.
+                * verify (bool): If False, ignores SSL certificates errors
+        """
+        credentials = Credentials(token, url, **kwargs)
+
+        self._append_account(credentials)
+
+    def list_accounts(self):
+        """List all accounts currently stored in the session.
 
         Returns:
-            dict: parsed configuration.
+            list[dict]: a list with information about the accounts currently
+                in the session.
         """
-        edited_config = {
-            'local': False
-        }
+        information = []
+        for provider in self._accounts.values():
+            information.append({
+                'token': provider.credentials.token,
+                'url': provider.credentials.url,
+            })
 
-        for key in config.keys():
-            new_key = _camel_case_to_snake_case(key)
-            if new_key not in ['id', 'serial_number', 'topology_id',
-                               'status']:
-                edited_config[new_key] = config[key]
+        return information
 
-        return edited_config
+    def load_accounts(self, **kwargs):
+        """Load IBMQ accounts found in the system, subject to optional filtering.
 
-    def _discover_remote_backends(self):
+        Automatically load the accounts found in the system. This method
+        looks for credentials in the following locations, in order, and
+        returns as soon as credentials are found:
+
+        1. in the `Qconfig.py` file in the current working directory.
+        2. in the environment variables.
+        3. in the `qiskitrc` configuration file
+
+        Raises:
+            IBMQAccountError: if attempting to load previously loaded accounts,
+                    or if no credentials can be found.
         """
-        Return the remote backends available.
+        # Special handling of the credentials filters.
+        credentials_filter = {}
+        for key in ['token', 'url', 'hub', 'group', 'project']:
+            if key in kwargs:
+                credentials_filter[key] = kwargs.pop(key)
+
+        for credentials in discover_credentials().values():
+            if self._match_all(credentials, credentials_filter):
+                self._append_account(credentials)
+
+        if not self._accounts:
+            raise IBMQAccountError('No IBMQ credentials found.')
+
+    def _append_account(self, credentials):
+        """Append an account with the specified credentials to the session.
+
+        Args:
+            credentials (Credentials): set of credentials.
 
         Returns:
-            dict[str:IBMQBackend]: a dict of the remote backend instances,
-                keyed by backend name.
+            IBMQSingleProvider: new single-account provider.
+
+        Raises:
+            IBMQAccountError: if the provider could not be appended.
         """
-        ret = {}
-        configs_list = self._api.available_backends()
-        for raw_config in configs_list:
-            config = self._parse_backend_configuration(raw_config)
-            ret[config['name']] = IBMQBackend(configuration=config, api=self._api)
+        # Check if duplicated credentials are already in use. By convention,
+        # we assume (hub, group, project) is always unique.
+        if credentials.unique_id() in self._accounts.keys():
+            raise IBMQAccountError('Credentials are already in use')
 
-        return ret
+        single_provider = IBMQSingleProvider(credentials, self)
+        self._accounts[credentials.unique_id()] = single_provider
 
-    def __eq__(self, other):
-        try:
-            equality = (self._token == other._token and self._url == other._url)
-        except AttributeError:
-            equality = False
-        return equality
+        return single_provider
+
+    def _match_all(self, obj, criteria):
+        """Return True if all items in criteria matches items in obj."""
+        return all(getattr(obj, key_, None) == value_ for
+                   key_, value_ in criteria.items())