+++ qiskit/backends/local/qasm_simulator_py.py
@@ -99,12 +99,9 @@ class QasmSimulatorPy(BaseBackend):
         'basis_gates': 'u1,u2,u3,cx,id,snapshot'
     }
 
-    def __init__(self, configuration=None):
-        """
-        Args:
-            configuration (dict): backend configuration
-        """
-        super().__init__(configuration or self.DEFAULT_CONFIGURATION.copy())
+    def __init__(self, configuration=None, provider=None):
+        super().__init__(configuration=configuration or self.DEFAULT_CONFIGURATION.copy(),
+                         provider=provider)
 
         self._local_random = random.Random()
 
@@ -269,11 +266,12 @@ class QasmSimulatorPy(BaseBackend):
         Returns:
             LocalJob: derived from BaseJob
         """
-        local_job = LocalJob(self._run_job, qobj)
+        job_id = str(uuid.uuid4())
+        local_job = LocalJob(self, job_id, self._run_job, qobj)
         local_job.submit()
         return local_job
 
-    def _run_job(self, qobj):
+    def _run_job(self, job_id, qobj):
         """Run circuits in qobj"""
         self._validate(qobj)
         result_list = []
@@ -284,7 +282,6 @@ class QasmSimulatorPy(BaseBackend):
         for circuit in qobj.experiments:
             result_list.append(self.run_circuit(circuit))
         end = time.time()
-        job_id = str(uuid.uuid4())
         result = {'backend': self._configuration['name'],
                   'id': qobj.qobj_id,
                   'job_id': job_id,