+++ test/python/test_simulator_interfaces.py
@@ -13,7 +13,7 @@ import unittest
 import qiskit
 import qiskit.extensions.simulator
 from qiskit.tools.qi.qi import state_fidelity
-from qiskit.wrapper import available_backends, register, execute, get_backend
+from qiskit.wrapper import execute
 from .common import requires_qe_access, QiskitTestCase, requires_cpp_simulator
 
 
@@ -44,7 +44,7 @@ class TestCrossSimulation(QiskitTestCase):
     @requires_qe_access
     def test_qasm(self, qe_token, qe_url):
         """counts from a GHZ state"""
-        register(qe_token, qe_url)
+        qiskit.IBMQ.use_account(qe_token, qe_url)
         qr = qiskit.QuantumRegister(3)
         cr = qiskit.ClassicalRegister(3)
         circuit = qiskit.QuantumCircuit(qr, cr)
@@ -96,7 +96,7 @@ class TestCrossSimulation(QiskitTestCase):
     @requires_qe_access
     def test_qasm_reset_measure(self, qe_token, qe_url):
         """counts from a qasm program with measure and reset in the middle"""
-        register(qe_token, qe_url)
+        qiskit.IBMQ.use_account(qe_token, qe_url)
         qr = qiskit.QuantumRegister(3)
         cr = qiskit.ClassicalRegister(3)
         circuit = qiskit.QuantumCircuit(qr, cr)