+++ test/python/test_wrapper.py
@@ -8,14 +8,12 @@
 
 """Tests for the wrapper functionality."""
 
-import logging
 import unittest
 
-from qiskit import compile as qcompile
 import qiskit.wrapper
 from qiskit import QISKitError
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
-from qiskit.backends.ibmq import IBMQProvider
+
 from qiskit.qobj import Qobj
 from qiskit.wrapper import registered_providers, execute
 from ._mockutils import DummyProvider
@@ -62,8 +60,9 @@ class TestWrapper(QiskitTestCase):
         """Test double registration of the same credentials."""
         qiskit.wrapper.register(qe_token, qe_url)
         initial_providers = registered_providers()
-        # Registering twice should give warning and add no providers.
-        qiskit.wrapper.register(qe_token, qe_url)
+
+        with self.assertRaises(QISKitError):
+            qiskit.wrapper.register(qe_token, qe_url)
         self.assertCountEqual(initial_providers, registered_providers())
 
     def test_register_bad_credentials(self):
@@ -73,52 +72,12 @@ class TestWrapper(QiskitTestCase):
             qiskit.wrapper.register('FAKE_TOKEN', 'http://unknown')
         self.assertEqual(initial_providers, registered_providers())
 
-    @requires_qe_access
-    def test_unregister(self, qe_token, qe_url):
-        """Test unregistering."""
-        initial_providers = registered_providers()
-        ibmqprovider = qiskit.wrapper.register(qe_token, qe_url)
-        self.assertCountEqual(initial_providers + [ibmqprovider],
-                              registered_providers())
-        qiskit.wrapper.unregister(ibmqprovider)
-        self.assertEqual(initial_providers, registered_providers())
-
-    @requires_qe_access
-    def test_unregister_non_existent(self, qe_token, qe_url):
-        """Test unregistering a non existent provider."""
-        initial_providers = registered_providers()
-        ibmqprovider = IBMQProvider(qe_token, qe_url)
-        with self.assertRaises(QISKitError):
-            qiskit.wrapper.unregister(ibmqprovider)
-        self.assertEqual(initial_providers, registered_providers())
-
-    def test_register_backend_name_conflicts(self):
-        """Test backend name conflicts when registering."""
-        class SecondDummyProvider(DummyProvider):
-            """
-            Subclass the DummyProvider so register treats them as different."""
-            pass
-
+    def test_register_third_party(self):
+        """Test registering third party backend."""
         dummy_provider = qiskit.wrapper.register(provider_class=DummyProvider)
-        initial_providers = registered_providers()
-        initial_backends = qiskit.wrapper.available_backends()
         dummy_backend = dummy_provider.get_backend('local_dummy_simulator')
-        with self.assertLogs(level=logging.WARNING) as logs:
-            second_dummy_provider = qiskit.wrapper.register(
-                provider_class=SecondDummyProvider)
 
-        # Check that one, and only one warning has been issued.
-        self.assertEqual(len(logs.records), 1)
-        # Check that the provider has been registered.
-        self.assertCountEqual(initial_providers + [second_dummy_provider],
-                              registered_providers())
-        # Check that no new backends have been added.
-        self.assertCountEqual(initial_backends,
-                              qiskit.wrapper.available_backends())
-
-        # Check the name of the backend still refers to the previous one.
-        self.assertEqual(dummy_backend,
-                         qiskit.wrapper.get_backend('local_dummy_simulator'))
+        self.assertTrue(dummy_backend)
 
     def test_local_execute_and_get_ran_qasm(self):
         """Check if the local backend return the ran qasm."""
@@ -146,7 +105,7 @@ class TestWrapper(QiskitTestCase):
     def test_qobj_to_circuits_single(self):
         """Check that qobj_to_circuits's result matches the qobj ini."""
         backend = 'local_qasm_simulator'
-        qobj_in = qcompile(self.circuit, backend, skip_transpiler=True)
+        qobj_in = qiskit.wrapper.compile(self.circuit, backend, skip_transpiler=True)
         out_circuit = qiskit.wrapper.qobj_to_circuits(qobj_in)
         self.assertEqual(out_circuit[0].qasm(), self.circuit.qasm())
 
@@ -162,7 +121,7 @@ class TestWrapper(QiskitTestCase):
         circuit_b.h(qreg2)
         circuit_b.measure(qreg1, creg1)
         circuit_b.measure(qreg2[0], creg2[1])
-        qobj = qcompile([self.circuit, circuit_b], backend, skip_transpiler=True)
+        qobj = qiskit.wrapper.compile([self.circuit, circuit_b], backend, skip_transpiler=True)
         qasm_list = [x.qasm() for x in qiskit.wrapper.qobj_to_circuits(qobj)]
         self.assertEqual(qasm_list, [self.circuit.qasm(), circuit_b.qasm()])
 
@@ -170,6 +129,16 @@ class TestWrapper(QiskitTestCase):
         """Verify that qobj_to_circuits returns None without QASM."""
         qobj = Qobj('abc123', {}, {}, {})
         self.assertIsNone(qiskit.wrapper.qobj_to_circuits(qobj))
+
+    def test_compact_flag(self):
+        """Test the compact flag for available_backends works"""
+        compact_names = qiskit.wrapper.available_backends()
+        expanded_names = qiskit.wrapper.available_backends(compact=False)
+        self.assertIn('local_qasm_simulator', compact_names)
+        self.assertIn('local_statevector_simulator', compact_names)
+        self.assertIn('local_unitary_simulator', compact_names)
+        self.assertIn('local_qasm_simulator_py', expanded_names)
+        self.assertIn('local_statevector_simulator_py', expanded_names)
 
 
 if __name__ == '__main__':